+++ pages/users/index.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import CardsContainer from '../components/CardsContainer';
-import { getUsers } from '../redux/slices/users';
-import { IState } from '../types/slices';
+import CardsContainer from '../../components/CardsContainer';
+import { getUsers } from '../../redux/slices/users';
+import { IState } from '../../types/slices';
 
 const Panel: React.FC = () => {
 	const dispatch = useDispatch();
